sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memoryhub"
version = "0.1.0"
description = "Centralized, governed memory for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Wes Jackson" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["ai", "agents", "memory", "mcp", "openshift", "kubernetes"]
dependencies = [
    "httpx>=0.28.1,<1.0",
    "pyjwt>=2.8",
    "pydantic>=2.0",
    "fastmcp>=2.11.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/rdwj/memory-hub"
Repository = "https://github.com/rdwj/memory-hub"
Issues = "https://github.com/rdwj/memory-hub/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/memoryhub"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
